app/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
python-dotenv==0.10.3
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
six==1.11.0
SQLAlchemy==1.3.8
Twisted==17.9.0
urllib3==1.22
uWSGI==2.0.18
Werkzeug==0.15.5
WTForms==2.2.1
